+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/). All notable c
 
 ### Fixed
 - [#276](https://github.com/Kashoo/synctos/issues/276): Date range validation is incorrect for dates between years 0 and 99
+- [#288](https://github.com/Kashoo/synctos/issues/288): Backticks are escaped even if sync function is not written to a file
 
 ## [2.2.1] - 2018-03-21
 ### Fixed